+++ src/TinyPNG/ResizeOperations/FitResizeOperation.cs
@@ -1,0 +1,7 @@
+﻿namespace TinyPng.ResizeOperations
+{
+    public class FitResizeOperation : ResizeOperation
+    {
+        public FitResizeOperation(int width, int height) : base(ResizeType.Fit, width, height) { }
+    }
+}
