+++ src/TinyPNG/ResizeOperations/ScaleHeightResizeOperation.cs
@@ -1,0 +1,7 @@
+﻿namespace TinyPng.ResizeOperations
+{
+    public class ScaleHeightResizeOperation : ResizeOperation
+    {
+        public ScaleHeightResizeOperation(int height) : base(ResizeType.Scale, 0, height) { }
+    }
+}
