+++ src/TinyPNG/ResizeOperations/ScaleWidthResizeOperation.cs
@@ -1,0 +1,7 @@
+﻿namespace TinyPng.ResizeOperations
+{
+    public class ScaleWidthResizeOperation : ResizeOperation
+    {
+        public ScaleWidthResizeOperation(int width) : base (ResizeType.Scale, width, 0) { }
+    }
+}
