+++ src/TinyPNG/TinyPNG.csproj
@@ -8,7 +8,7 @@
     <TargetFrameworks>net45;netstandard1.3</TargetFrameworks>
     <AssemblyName>TinyPNG</AssemblyName>
     <PackageId>TinyPNG</PackageId>
-    <PackageTags>tinypng;images;compression</PackageTags>
+    <PackageTags>tinypng;images;compression;jpg;png</PackageTags>
     <PackageReleaseNotes>Support for netstandard 1.3</PackageReleaseNotes>
     <PackageIconUrl>https://raw.githubusercontent.com/ctolkien/TinyPNG/master/icon.png</PackageIconUrl>
     <PackageProjectUrl>https://github.com/ctolkien/TinyPNG</PackageProjectUrl>
