+++ tests/TinyPng.Tests/TinyPngTests.cs
@@ -1,99 +1,12 @@
-﻿using Newtonsoft.Json;
-using System;
+﻿using System;
 using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
+using TinyPng.ResizeOperations;
 
 namespace TinyPng.Tests
 {
-    static class Extensions
-    {
-        public static FakeResponseHandler Compress(this FakeResponseHandler fakeResponse)
-        {
-
-            var content = new TinyPngApiResult();
-            content.Input = new TinyPngApiInput
-            {
-                Size = 18031,
-                Type = "image/jpeg"
-            };
-            content.Output = new TinyPngApiOutput
-            {
-                Width = 400,
-                Height = 400,
-                Size = 16646,
-                Type = "image/jpeg",
-                Ratio = 0.9232f,
-                Url = "https://api.tinify.com/output"
-            };
-
-            var compressResponseMessage = new HttpResponseMessage
-            {
-                StatusCode = System.Net.HttpStatusCode.Created,
-                Content = new StringContent(JsonConvert.SerializeObject(content)),
-            };
-            compressResponseMessage.Headers.Location = new Uri("https://api.tinify.com/output");
-            compressResponseMessage.Headers.Add("Compression-Count", "99");
-
-            fakeResponse.AddFakePostResponse(new Uri("https://api.tinify.com/shrink"), compressResponseMessage);
-            return fakeResponse;
-        }
-
-        public static FakeResponseHandler CompressAndFail(this FakeResponseHandler fakeResponse)
-        {
-            var errorApiObject = new TinyPngApiException(400, "reason", "title", "message");
-
-            var compressResponseMessage = new HttpResponseMessage
-            {
-                StatusCode = System.Net.HttpStatusCode.BadRequest,
-                Content = new StringContent(JsonConvert.SerializeObject(errorApiObject))
-            };
-            fakeResponse.AddFakePostResponse(new Uri("https://api.tinify.com/shrink"), compressResponseMessage);
-            return fakeResponse;
-        }
-
-        public static FakeResponseHandler Download(this FakeResponseHandler fakeResponse)
-        {
-            var compressedCatStream = File.OpenRead(TinyPngTests.CompressedCat);
-            var outputResponseMessage = new HttpResponseMessage
-            {
-                Content = new StreamContent(compressedCatStream),
-                StatusCode = System.Net.HttpStatusCode.OK
-            };
-
-            fakeResponse.AddFakeGetResponse(new Uri("https://api.tinify.com/output"), outputResponseMessage);
-            return fakeResponse;
-        }
-
-        public static FakeResponseHandler Resize(this FakeResponseHandler fakeResponse)
-        {
-            var resizedCatStream = File.OpenRead(TinyPngTests.ResizedCat);
-            var resizeMessage = new HttpResponseMessage
-            {
-                StatusCode = System.Net.HttpStatusCode.OK,
-                Content = new StreamContent(resizedCatStream)
-            };
-            resizeMessage.Headers.Add("Image-Width", "150");
-            resizeMessage.Headers.Add("Image-Height", "150");
-
-            fakeResponse.AddFakePostResponse(new Uri("https://api.tinify.com/output"), resizeMessage);
-            return fakeResponse;
-        }
-
-        public static FakeResponseHandler S3(this FakeResponseHandler fakeResponse)
-        {
-            var amazonMessage = new HttpResponseMessage
-            {
-                StatusCode = System.Net.HttpStatusCode.OK
-            };
-            amazonMessage.Headers.Add("Location", "https://s3-ap-southeast-2.amazonaws.com/tinypng-test-bucket/path.jpg");
-
-            fakeResponse.AddFakePostResponse(new Uri("https://api.tinify.com/output"), amazonMessage);
-            return fakeResponse;
-        }
-    }
-
     public class TinyPngTests
     {
         const string apiKey = "lolwat";
@@ -101,7 +14,6 @@ namespace TinyPng.Tests
         internal const string Cat = "Resources/cat.jpg";
         internal const string CompressedCat = "Resources/compressedcat.jpg";
         internal const string ResizedCat = "Resources/resizedcat.jpg";
-
 
         [Fact]
         public void TinyPngClientThrowsWhenNoApiKeySupplied()
@@ -117,7 +29,6 @@ namespace TinyPng.Tests
             Assert.Throws<ArgumentNullException>(() => new TinyPngClient("apiKey", null));
             Assert.Throws<ArgumentNullException>(() => new TinyPngClient(null, new AmazonS3Configuration("a", "b", "c", "d")));
         }
-
 
         [Fact]
         public async Task Compression()
@@ -196,8 +107,6 @@ namespace TinyPng.Tests
             Assert.Equal(16646, downloadResult.Length);
         }
 
-
-
         [Fact]
         public async Task ResizingOperationThrows()
         {
@@ -211,7 +120,6 @@ namespace TinyPng.Tests
             await Assert.ThrowsAsync<ArgumentNullException>(async () => await pngx.Compress(Cat).Resize(null));
             await Assert.ThrowsAsync<ArgumentOutOfRangeException>(async () => await pngx.Compress(Cat).Resize(0, 150));
             await Assert.ThrowsAsync<ArgumentOutOfRangeException>(async () => await pngx.Compress(Cat).Resize(150, 0));
-
         }
 
         [Fact]
@@ -238,7 +146,6 @@ namespace TinyPng.Tests
             var resizedImageByteData = await pngx.Compress(Cat).Resize(new ScaleHeightResizeOperation(150)).GetImageByteData();
 
             Assert.Equal(5970, resizedImageByteData.Length);
-
         }
 
         [Fact]
@@ -249,11 +156,9 @@ namespace TinyPng.Tests
                 .Compress()
                 .Resize());
 
-
             var resizedImageByteData = await pngx.Compress(Cat).Resize(new ScaleWidthResizeOperation(150)).GetImageByteData();
 
             Assert.Equal(5970, resizedImageByteData.Length);
-
         }
 
         [Fact]
@@ -269,7 +174,6 @@ namespace TinyPng.Tests
             var resizedImageByteData = await pngx.Compress(Cat).Resize(new FitResizeOperation(150, 150)).GetImageByteData();
 
             Assert.Equal(5970, resizedImageByteData.Length);
-
         }
 
         [Fact]
@@ -283,7 +187,6 @@ namespace TinyPng.Tests
             var resizedImageByteData = await pngx.Compress(Cat).Resize(new CoverResizeOperation(150, 150)).GetImageByteData();
 
             Assert.Equal(5970, resizedImageByteData.Length);
-
         }
 
         [Fact]
@@ -297,7 +200,6 @@ namespace TinyPng.Tests
             await Assert.ThrowsAsync<ArgumentException>(async () => await pngx.Compress(Cat).Resize(new CoverResizeOperation(0, 150)));
             await Assert.ThrowsAsync<ArgumentException>(async () => await pngx.Compress(Cat).Resize(new CoverResizeOperation(150, 0)));
         }
-
 
         [Fact]
         public void CompressAndStoreToS3ShouldThrowIfNoApiKeyProvided()
@@ -354,7 +256,6 @@ namespace TinyPng.Tests
             
             //S3 configuration has not been set
             await Assert.ThrowsAsync<InvalidOperationException>(async () => await pngx.SaveCompressedImageToAmazonS3(result, path: string.Empty));
-            
         }
 
         [Fact]
